+++ catalog/app/components/JsonEditor/JsonEditor.tsx
@@ -9,7 +9,7 @@ import { EMPTY_SCHEMA, JsonSchema } from 'utils/json-schema'
 import illustrationEnterValues from './enter-values.webm'
 import illustrationObjectExpand from './object-expand.webm'
 import Column from './Column'
-import State from './State'
+import State, { StateRenderProps } from './State'
 import { JsonValue, RowData, ValidationErrors } from './constants'
 
 interface EmptyStateCaseProps {
@@ -85,7 +85,7 @@ function EmptyState({ className, noValue, notExpanded }: EmptyStateProps) {
 
 interface ColumnData {
   items: RowData[]
-  parent: JsonValue
+  parent?: JsonValue
 }
 
 function shouldSqueezeColumn(columnIndex: number, columns: ColumnData[]) {
@@ -161,20 +161,11 @@ const useStyles = M.makeStyles<any, { multiColumned: boolean }>((t) => ({
   },
 }))
 
-interface JsonEditorProps {
-  addRow: (path: string[], key: string | number, value: JsonValue) => JsonValue
-  changeValue: (path: string[], id: 'key' | 'value', value: JsonValue) => JsonValue
+interface JsonEditorProps extends StateRenderProps {
   className?: string
-  columns: ColumnData[]
   disabled?: boolean
-  fieldPath: string[]
-  jsonDict: Record<string, JsonValue>
-  menuFieldPath: string[]
   multiColumned: boolean
   onChange: (value: JsonValue) => JsonValue
-  removeField: (path: string[]) => JsonValue
-  setFieldPath: (path: string[]) => void
-  setMenuFieldPath: (path: string[]) => void
 }
 
 const JsonEditor = React.forwardRef<HTMLDivElement, JsonEditorProps>(function JsonEditor(
@@ -192,6 +183,7 @@ const JsonEditor = React.forwardRef<HTMLDivElement, JsonEditorProps>(function Js
     removeField,
     setFieldPath,
     setMenuFieldPath,
+    transformer,
   },
   ref,
 ) {
@@ -221,6 +213,13 @@ const JsonEditor = React.forwardRef<HTMLDivElement, JsonEditorProps>(function Js
       if (newData) onChange(newData)
     },
     [changeValue, onChange],
+  )
+
+  const handleToolbar = React.useCallback(
+    (transform) => {
+      onChange(transformer(transform))
+    },
+    [onChange, transformer],
   )
 
   if (!columns.length) throw new Error('No column data')
@@ -256,6 +255,7 @@ const JsonEditor = React.forwardRef<HTMLDivElement, JsonEditorProps>(function Js
               onContextMenu={setMenuFieldPath}
               onExpand={setFieldPath}
               onRemove={handleRowRemove}
+              onToolbar={handleToolbar}
             />
           )
         })}
@@ -270,25 +270,6 @@ const JsonEditor = React.forwardRef<HTMLDivElement, JsonEditorProps>(function Js
     </div>
   )
 })
-
-interface StateRenderProps {
-  addRow: (path: string[], key: string | number, value: JsonValue) => JsonValue
-  changeValue: (
-    path: string[],
-    key: 'key' | 'value',
-    value: JsonValue | string,
-  ) => JsonValue
-  columns: {
-    items: RowData[]
-    parent: JsonValue
-  }[]
-  fieldPath: string[]
-  jsonDict: Record<string, JsonValue>
-  removeField: (path: string[]) => JsonValue
-  setFieldPath: (path: string[]) => void
-  menuFieldPath: string[]
-  setMenuFieldPath: (path: string[]) => void
-}
 
 interface JsonEditorWrapperProps {
   className?: string