+++ catalog/app/containers/Bucket/PackageDialog/PackageDialog.tsx
@@ -197,6 +197,7 @@ const useFieldInputStyles = M.makeStyles({
   },
 })
 
+// TODO: re-use components/Form/TextField
 export function Field({
   error,
   helperText,