+++ catalog/setup-jest.ts
@@ -1,0 +1,14 @@
+;(window as any).QUILT_CATALOG_CONFIG = {
+  "apiGatewayEndpoint": "",
+  "binaryApiGatewayEndpoint": "",
+  "alwaysRequiresAuth": true,
+  "s3Proxy": "",
+  "registryUrl": "",
+  "ssoAuth": "DISABLED",
+  "ssoProviders": "",
+  "analyticsBucket": "",
+  "serviceBucket": "",
+  "mode": "PRODUCT",
+  "mixpanelToken": "",
+  "passwordAuth": "ENABLED",
+}